weight now is float

+++ entities.py
@@ -4,4 +4,4 @@ from dataclasses import dataclass
 class Habit:
     name: str
     minutes: int
-    weight: int
+    weight: float
